fix(utils): move metrics
