+++ packages/shared/lib/clients/sync.client.ts
@@ -26,11 +26,11 @@ import connectionService from '../services/connection.service.js';
 import configService from '../services/config.service.js';
 import { deleteRecordsBySyncId } from '../services/sync/data/records.service.js';
 import { createSync, clearLastSyncDate } from '../services/sync/sync.service.js';
-import telemetry, { LogTypes, MetricTypes } from '../utils/telemetry.js';
+import telemetry, { LogTypes } from '../utils/telemetry.js';
 import errorManager, { ErrorSourceEnum } from '../utils/error.manager.js';
 import { NangoError } from '../utils/error.js';
 import type { RunnerOutput } from '../models/Runner.js';
-import { isTest, isProd, getLogger } from '@nangohq/utils';
+import { isTest, isProd, getLogger, metrics } from '@nangohq/utils';
 import { isErr, resultOk, type Result, resultErr } from '@nangohq/utils';
 
 const logger = getLogger('Sync.Client');
@@ -658,7 +658,7 @@ class SyncClient {
         } finally {
             const endTime = Date.now();
             const totalRunTime = (endTime - startTime) / 1000;
-            telemetry.duration(MetricTypes.ACTION_TRACK_RUNTIME, totalRunTime);
+            metrics.duration(metrics.Types.ACTION_TRACK_RUNTIME, totalRunTime);
         }
     }
