+++ packages/utils/lib/index.ts
@@ -6,3 +6,4 @@ export * from './logger.js';
 export * from './id.js';
 export * from './result.js';
 export * from './encryption.js';
+export * as metrics from './telemetry/metrics.js';
